backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
boto3==1.34.74
python-dotenv==1.0.1
pydantic==2.6.4
httpx==0.27.0
